public api fixed
