+++ clickhouse/client.cpp
@@ -2,6 +2,7 @@
 #include "clickhouse/version.h"
 #include "protocol.h"
 
+#include "base/compressed.h"
 #include "base/socket.h"
 #include "base/wire_format.h"
