+++ clickhouse/client.h
@@ -3,8 +3,6 @@
 #include "query.h"
 #include "exceptions.h"
 
-#include "base/compressed.h"
-
 #include "columns/array.h"
 #include "columns/date.h"
 #include "columns/decimal.h"
@@ -39,6 +37,13 @@ struct ServerInfo {
     uint64_t    version_minor;
     uint64_t    version_patch;
     uint64_t    revision;
+};
+
+/// Methods of block compression.
+enum class CompressionMethod : int8_t {
+    None = -1,
+    LZ4  = 1,
+    ZSTD = 2,
 };
 
 struct Endpoint {
